chore(...) : fragment 상태 유지
