+++ app/src/main/java/com/iron/espresso/presentation/home/study/list/NewListFragment.kt
@@ -2,14 +2,42 @@ package com.iron.espresso.presentation.home.study.list
 
 import android.os.Bundle
 import android.view.View
+import android.widget.Toast
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.Observer
 import com.iron.espresso.R
 import com.iron.espresso.base.BaseFragment
 import com.iron.espresso.databinding.FragmentNewListBinding
+import com.iron.espresso.presentation.home.study.StudyListViewModel
+import com.iron.espresso.presentation.home.study.adapter.StudyListAdapter
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class NewListFragment : BaseFragment<FragmentNewListBinding>(R.layout.fragment_new_list) {
+
+    private val viewModel by viewModels<StudyListViewModel>()
+    private val studyListAdapter = StudyListAdapter()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        binding.studyList.adapter = studyListAdapter
+
+        viewModel.getStudyList("android", "new")
+
+        viewModel.studyList.observe(viewLifecycleOwner, Observer { studyList ->
+
+            studyListAdapter.apply {
+                setItemList(studyList)
+                itemClickListener = { title ->
+                    Toast.makeText(
+                        context,
+                        "onClick title: $title",
+                        Toast.LENGTH_SHORT
+                    ).show()
+                }
+            }
+        })
     }
 
     companion object {
