+++ app/src/main/res/layout/activity_study_list.xml
@@ -26,16 +26,10 @@
 
         </com.google.android.material.tabs.TabLayout>
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/study_list"
+        <FrameLayout
+            android:id="@+id/study_list_container"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="16dp"
-            android:layout_marginEnd="8dp"
-            android:overScrollMode="never"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            tools:listitem="@layout/item_study_list" />
+            android:layout_height="match_parent" />
 
     </LinearLayout>
 </layout>
